notification-service/requirements.txt
Flask
neo4j
python-dotenv
pyjwt
